Test: frontend middleware
